requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.2
pytz==2022.5
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.20.4
typing_extensions==4.4.0
Werkzeug==2.2.2
